Inherit TCPSocket from Stream class

+++ pic32/libraries/DEIPcK/TCPSocket.cpp
@@ -186,7 +186,7 @@ void TCPSocket::discardReadBuffer(void)
 **      This call is safe to make without checking the connection status.
 **
 */
-size_t TCPSocket::available(void)
+int TCPSocket::available(void)
 {
     return(TCPAvailable(&_socket, NULL));
 }
